pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "promptrix"
version = "0.0.2"
authors = [
  { name="Steven Ickman", email="author@example.com" },
  { name="Bruce DAmbrosio", email="bruce.dambrosio@gmail.com" },
]

description = "Promptrix. A prompt layout manager for LLMs"

readme = "README.md"

requires-python = ">=3.8"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    #'dataclasses',
    #'enum',
    #'json',
    "requests",
    #"types",
    #"typing",
    #"asyncio",
    "tiktoken",
    "pyyaml",
    'importlib-metadata; python_version<"3.9"',
]


[project.urls]
"Homepage" = "https://github.com/Stevenic/promptrix-py"
"Bug Tracker" = "https://github.com/Stevenic/promptrix-py/issues"
